MAGE-1056: adding a asbtract class for multiple stores scenario + starting of the category class
